# openapi.yaml
openapi: 3.0.3
info:
  title: Armada Convoy API
  description: REST API for the Armada convoy and social app
  version: 1.0.0
  contact:
    name: Armada Team
    email: support@armada.app
servers:
  - url: https://convoy-api.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /threads:
    get:
      summary: Get user threads
      description: Get threads for the current user
      parameters:
        - name: limit
          in: query
          description: Number of threads to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of threads to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Threads retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      threads:
                        type: array
                        items:
                          $ref: '#/components/schemas/Thread'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      summary: Create thread
      description: Create a new chat thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - participants
              properties:
                participants:
                  type: array
                  items:
                    type: string
                  minItems: 2
                  maxItems: 50
                  description: Array of participant user IDs
                title:
                  type: string
                  maxLength: 100
                  description: Thread title (for group chats)
                isGroup:
                  type: boolean
                  default: false
                  description: Whether this is a group thread
      responses:
        '201':
          description: Thread created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Thread'
  /threads/{id}/messages:
    get:
      summary: Get thread messages
      description: Get messages for a specific thread
      parameters:
        - name: id
          in: path
          required: true
          description: Thread ID
          schema:
            type: string
        - name: cursor
          in: query
          description: Cursor for pagination
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Number of messages to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      messages:
                        type: array
                        items:
                          $ref: '#/components/schemas/Message'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      summary: Send message
      description: Send a message to a thread
      parameters:
        - name: id
          in: path
          required: true
          description: Thread ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  maxLength: 2000
                  description: Message text
                media:
                  $ref: '#/components/schemas/MessageMedia'
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Message'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
        avatarUrl:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
        bio:
          type: string
          example: Adventure enthusiast and convoy leader
        settings:
          $ref: '#/components/schemas/UserSettings'
        stats:
          $ref: '#/components/schemas/UserStats'
        friends:
          type: array
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserSettings:
      type: object
      properties:
        showConvoys:
          type: string
          enum: [everyone, friends, private]
          example: friends
        showStats:
          type: string
          enum: [everyone, friends, private]
          example: friends
        showProfile:
          type: string
          enum: [everyone, friends, private]
          example: everyone
        notifications:
          type: object
          properties:
            convoyInvites:
              type: boolean
              example: true
            friendRequests:
              type: boolean
              example: true
            messages:
              type: boolean
              example: true
            achievements:
              type: boolean
              example: true
    UserStats:
      type: object
      properties:
        convoysCreated:
          type: integer
          example: 12
        convoysJoined:
          type: integer
          example: 45
        totalMiles:
          type: number
          example: 1250.5
        friendsCount:
          type: integer
          example: 23
...
    Thread:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        participants:
          type: array
          items:
            $ref: '#/components/schemas/User'
        lastMessageAt:
          type: string
          format: date-time
        unreadCounts:
          type: object
          additionalProperties:
            type: integer
        isGroup:
          type: boolean
          example: false
        title:
          type: string
          example: Road Trip Planning
        convoyId:
          type: string
          example: 507f1f77bcf86cd799439011
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        threadId:
          type: string
          example: 507f1f77bcf86cd799439011
        senderId:
          type: string
          example: 507f1f77bcf86cd799439011
        text:
          type: string
          example: Let's meet at the gas station at 9 AM
        media:
          $ref: '#/components/schemas/MessageMedia'
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/MessageReaction'
        editedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MessageMedia:
      type: object
      properties:
        url:
          type: string
          format: uri
          example: https://example.com/media.jpg
        type:
          type: string
          enum: [image, video, audio]
          example: image
        width:
          type: integer
          example: 1920
        height:
          type: integer
          example: 1080
        duration:
          type: number
          example: 120.5
        filename:
          type: string
          example: photo.jpg
        size:
          type: integer
          example: 2048576
    MessageReaction:
      type: object
      properties:
        by:
          type: string
          example: 507f1f77bcf86cd799439011
        emoji:
          type: string
          example: 👍
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        total:
          type: integer
          example: 100
        hasMore:
          type: boolean
          example: true
...